requirements.txt
Adafruit-DHT==1.4.0
PyYAML==3.13
RPi.GPIO==0.6.5
